requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
astroid==2.5.6
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.2.6
attrs==20.3.0
bayesian-optimization==1.2.0
beautifulsoup4==4.9.3
blessings==1.7
bsuite @ git+git://github.com/deepmind/bsuite.git@a07485f497b72669f1058639fa806b6127c4c6a9
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
contextvars==2.4
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
descartes==1.1.0
dm-acme==0.2.0
dm-control==0.0.364896371
dm-env==1.4
dm-reverb==0.2.0
dm-sonnet==2.0.0
dm-tree==0.1.6
filelock==3.0.12
flatbuffers==1.12
frozendict==1.2
future==0.18.2
gast==0.3.3
glfw==2.1.0
google==3.0.0
google-api-core==1.26.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
hiredis==2.0.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.15
importlab==0.6.1
importlib-metadata==4.0.1
iniconfig==1.1.1
isort==5.8.0
jax==0.2.12
jaxlib==0.1.65
joblib==1.0.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labmaze==1.0.4
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.4
mccabe==0.6.1
mizani==0.7.3
mpmath==1.2.1
msgpack==1.0.2
multidict==5.1.0
networkx==2.5.1
ninja==1.10.0.post2
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.12
opencensus-context==0.1.2
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
palettable==3.3.0
pandas==1.1.5
patsy==0.5.1
Pillow==7.2.0
plotnine==0.8.0
pluggy==0.13.1
portpicker==1.3.1
prometheus-client==0.10.1
protobuf==3.15.8
psutil==5.8.0
py==1.10.0
py-spy==0.3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pylint==2.8.2
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
pytype==2021.4.26
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
ray==1.3.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
soupsieve==2.2.1
statsmodels==0.12.2
sympy==1.8
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.2
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
tqdm==4.60.0
trfl==1.1.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
